Fixes bug
