+++ src/controllers/users.ts
@@ -1,9 +1,11 @@
 import db from "../database/database";
 import { NewUserData } from "../util/objects";
 import { hash } from "bcrypt";
+
 const { Op } = require("sequelize");
 const sqlz = require("sequelize").Sequelize;
 const user = require("../database/models/users")(db);
+const businessUnits = require("../database/models/payments_periods")(db);
 
 function getOrder(order: string, by: string) {
     switch (order) {
@@ -181,4 +183,30 @@ export async function pseudoDeleteUser(id: number, businessUnits?: Array<number>
     }
 
     return { successful: true, found: result[0] === 1 };
+}
+
+// Alphanumeric
+export async function generatePassword(length: number) {
+    var result = '';
+    var characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+    var charactersLength = characters.length;
+
+    for ( var i = 0; i < length; i++ ) {
+        result += characters.charAt(Math.floor(Math.random() * charactersLength));
+    }
+    
+    return result;
+}
+
+export async function getPaymentPeriods() {
+    let businessUnitData;
+
+    try {
+        businessUnitData = await businessUnits.findAll();
+        
+    } catch (error) {
+        return false;
+    }
+
+    return businessUnitData;
 }
